// ph-healthcare-frontent/src/generated/prisma/enums.ts
export const AppointmentStatus = {
  PENDING: 'PENDING',
  CONFIRMED: 'CONFIRMED',
  CHECKED_IN: 'CHECKED_IN',
  IN_PROGRESS: 'IN_PROGRESS',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED',
  NO_SHOW: 'NO_SHOW',
  RESCHEDULED: 'RESCHEDULED'
} as const

export type AppointmentStatus = (typeof AppointmentStatus)[keyof typeof AppointmentStatus]


export const AppointmentType = {
  IN_PERSON: 'IN_PERSON',
  ONLINE: 'ONLINE',
  FOLLOW_UP: 'FOLLOW_UP',
  EMERGENCY: 'EMERGENCY'
} as const

export type AppointmentType = (typeof AppointmentType)[keyof typeof AppointmentType]


export const CancellationReason = {
  PATIENT_REQUEST: 'PATIENT_REQUEST',
  DOCTOR_UNAVAILABLE: 'DOCTOR_UNAVAILABLE',
  CENTER_CLOSED: 'CENTER_CLOSED',
  NO_SHOW: 'NO_SHOW',
  OTHER: 'OTHER'
} as const

export type CancellationReason = (typeof CancellationReason)[keyof typeof CancellationReason]


export const AppointmentPaymentStatus = {
  NOT_REQUIRED: 'NOT_REQUIRED',
  UNPAID: 'UNPAID',
  PARTIAL: 'PARTIAL',
  PAID: 'PAID',
  REFUNDED: 'REFUNDED'
} as const

export type AppointmentPaymentStatus = (typeof AppointmentPaymentStatus)[keyof typeof AppointmentPaymentStatus]